refactor: move post's repository interface to separate file
