+++ server/src/posts/domain/post.ts
@@ -1,4 +1,3 @@
-import { DomainPaginationResult } from '../../common/interfaces/domainPaginationResult';
 import { IncorrectInputError } from '../../common/errors/errors';
 import { PostAuthor } from './postAuthor';
 
@@ -24,11 +23,4 @@ export function newPost(payload: Partial<Post>): Post {
     createdAt: new Date(),
     updatedAt: new Date(),
   };
-}
-
-export interface PostRepository {
-  getPost(postId: string): Promise<Post>;
-  listPosts(limit: number, page: number): Promise<DomainPaginationResult<Post>>;
-  createPost(post: Post): Promise<void>;
-  generateId(): string;
 }
